requirements.txt
pygame==2.5.2
pywebview==4.4.1
